# doc/swagger/AoL-Teacher.yaml
openapi: 3.0.0
info:
  title: AoL Teacher Interface
  description: |-
    AoL Teacher API specification.

  contact:
    email: sudhagar@abovecloud9.ai
  version: 0.0.4
...
servers:
  - url: https://api.abovecloud9.ai/api/v1 # placeholder
...
paths:
  /courses:
    put:
      tags:
        - course
      summary: Updates an existing course
      description: Update an existing course by course id
      operationId: updateCourse
      requestBody:
        description: Update an existent course
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: []
    post:
      tags:
        - course
      summary: Adds a new course
      description: Add a new course
      operationId: addCourse
      requestBody:
        description: Create a new course
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Course'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '405':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: []
...
  /courses/{courseId}:
    get:
      tags:
        - course
      summary: Finds course by ID
      description: Returns a single course
      operationId: getCourseById
      parameters:
        - name: courseId
          in: path
          description: ID of course to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: []
    delete:
      tags:
        - course
      summary: Deletes a course
      description: delete a course
      operationId: deleteCourse
      parameters:
        - name: courseId
          in: path
          description: Course id to delete
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid Course id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: []
...
  /accounts:
    get:
      tags:
        - account
      summary: Finds user account using search string
      description: Retrives user accounts that match the given search string. This should be paginated.
      operationId: searchAccounts
      parameters:
        - name: q
          in: query
          description: Account names that need to be considered for filter
          required: false
          schema:
            type: string
        - name: type
          in: query
          description: Account type
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number (for pagination)
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: Number of items in the response
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearer: []
...
components:
  schemas:
...
    DateTime:
      type: object
      properties:
        date:
          type: string
          format: date
        startTime:
          type: string
          format: time
        endTime:
          type: string
          format: time
    Timezone:
      type: string
      description: Supported timezones
      example: EST
      enum:
        - EST
        - CST
        - MST
        - PST
    Course:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10000000
        name:
          type: string
        organizer:
          type: array
          items:
            type: integer
            format: int64
            example: 10000000
        contact:
          type: array
          items:
            type: integer
            format: int64
            example: 10000000
        teacher:
          type: array
          items:
            type: integer
            format: int64
            example: 10000000
        notes:
          type: string
          description: Course notes
        status:
          type: string
          description: Course status
          example: open
          enum:
            - open
            - expense-submitted
            - expense-declined
            - closed
            - active
            - declined
            - submitted
            - canceled
            - inactive
        maxAttendees:
          type: integer
          format: int32
          example: 10
        dates:
          type: array
          items:
            $ref: '#/components/schemas/DateTime'
        timezone:
          $ref: '#/components/schemas/Timezone'
        address:
          $ref: '#/components/schemas/Address'
        centerId:
          type: integer
          format: int64
          example: 10000000
        productId:
          type: integer
          format: int64
          example: 20000000
        mode:
          type: string
          description: Course delivery mode
          example: in-person
          enum:
            - in-person
            - online
        notify:
          type: array
          items:
            type: integer
            format: int64
            example: 10000000
        numAttendees: # Cannot be more than maxAttendees
          type: integer
          format: int32
          example: 10
        attendees:
          type: array
          items:
            type: integer
            format: int64
            example: 10000000
        autoApprove:
          type: boolean
...
    Address:
      type: object
      properties:
        street:
          type: string
          example: 437 Lytton
        city:
          type: string
          example: Palo Alto
        state:
          type: string
          example: CA
        zip:
          type: string
          example: '94301'
        country:
          type: string
          example: USA
...
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 30000000
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        phone:
          type: string
          example: '********'
        type:
          type: string
          description: User Type
          example: teacher
          enum:
            - assistant-teacher
            - member
            - organizer
            - student
            - teacher
            - user
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: integer
              format: int32
            message:
              type: string
            details:
              type: string
            traceId:
              type: string
...
  securitySchemes:
    bearer:
      type: apiKey
      name: Bearer
      in: header
    sample_auth_unused:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://abovecloud9.ai/oauth/authorize
          scopes:
            write:courses: modify courses
            read:courses: read courses